requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
pytz==2023.3.post1
redgreenunittest==0.1.1
sqlparse==0.4.4
